# weekly-api-server/src/gen/java/org/weekly/api/TaskLabelsApi.java
package org.weekly.api;

import org.weekly.model.TaskLabel;

import java.io.InputStream;
import java.io.OutputStream;
import java.util.List;
import java.util.Map;
import javax.ws.rs.*;
import javax.ws.rs.core.Response;
import javax.ws.rs.core.MediaType;
import org.apache.cxf.jaxrs.ext.multipart.*;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiResponses;
import io.swagger.annotations.ApiResponse;
import io.swagger.jaxrs.PATCH;
import javax.validation.constraints.*;
import javax.validation.Valid;

/**
 * Weekly
 *
 * <p>No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 */
@Path("/")
@Api(value = "/", description = "")
public interface TaskLabelsApi  {

    /**
     * Get all task labels.
     *
     */
    @GET
    @Path("/task/labels")
    @Produces({ "application/json" })
    @ApiOperation(value = "Get all task labels.", tags={ "task-labels" })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = ".", response = TaskLabel.class) })
    public TaskLabel getTaskLabels();
}